♻️[Refactor] searchBox 리팩토링
