+++ src/components/input/search-box/SearchBox.stories.tsx
@@ -1,9 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import SearchBox from './SearchBox';
-import PencilIcon from '../../../public/icons/PencilIcon';
 
 const meta = {
-  title: 'Components/SearchBox',
+  title: 'Components/Input/SearchBox',
   component: SearchBox,
   parameters: {
     layout: 'centered',
@@ -18,16 +17,5 @@ export const Default: Story = {
   args: {
     value: '',
     onChange: (e) => console.log('Search value:', e.target.value),
-    placeholder: '검색어를 입력해주세요',
-  },
-};
-
-export const MessageInput: Story = {
-  args: {
-    value: '',
-    onChange: (e) => console.log('Message:', e.target.value),
-    placeholder: '메세지를 입력해주세요',
-    icon: <PencilIcon />,
-    className: 'bg-gray-light-02 border-none',
   },
 };
